Use createEvent to automatically add gmeet link (#2)
Version 1 used createEvent. In version 2, I replaced it with Events.insert API. Reason was that if we use createEvent, then the first invite will have the guest list visible. We can hide the guest list only after creating the event. But Events.insert has a bigger issue, which is that it does not automatically create gmeet links. So, I reverted to using createEvent. By the way, createEvent will create gmeet link only if there is at least 1 attendee. So, trying to create the event with no attendees and then adding them later will not work.